remove scaleway-cli package with borken deps
github.com/scaleway/scaleway-cli/pkg/utils can be imported because it
depends on github.com/moul/gotty-client which can't build on solaris.

This requires removing the
github.com/scaleway/scaleway-cli/pkg/api/helpers.go file which imports
the utils package, and we don't use. Not sure how we'll handle this if
the structure of the scaleway package changes and we can't break this
import any longer.